Add global options
