+++ oriented_matroids/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import absolute_import
 from .oriented_matroid import OrientedMatroid
-from .oriented_matroids_category import OrientedMatroids
+from .abstract_oriented_matroid import AbstractOrientedMatroid
 
 from sage.misc.lazy_import import lazy_import
 #lazy_import('sage.matroids.oriented_matroids',
